Add response to practice
Comfort: 5
Clarity: 3

It was really hard to figure out what this question was asking and to implement it based on the code that was provided.  I didn't use self.stretch and I think it would have been easier had I started from scratch.  I used the Array, String, and Number Ruby documentation